feat: Refactor Go API for Vercel serverless deployment
This commit refactors the Go application to be deployable as serverless
functions on Vercel.

Key changes:
- I restructured the project with an `api/` directory for Vercel handlers
  (e.g., `api/predict.go`) and `pkg/appcore/` for shared logic,
  types, and client initializations.
- `api/predict.go` contains the `PredictHandler` for the `/predict` endpoint.
- I ensured client initialization (Supabase, Hugging Face token) is handled using
  `sync.Once` within `pkg/appcore` suitable for serverless environments.
- I added `vercel.json` to configure Vercel builds (`@vercel/go` builder)
  and routing for the `/predict` endpoint.
- The root `main.go` no longer starts an HTTP server; this is handled by Vercel.
- I created a new Docker entrypoint `cmd/server/main.go` to allow
  the Dockerfile to remain functional for standalone server deployments.
- I comprehensively updated `README.md` with:
  - New project structure.
  - Instructions for local development with `vercel dev` (using `.env`).
  - Step-by-step guide for deploying to Vercel, including setting
    environment variables in the Vercel dashboard.
  - Clarification on the role of the `Dockerfile` for alternative deployments.
  - Adjustments to Karate testing instructions for different environments.
